requirements.txt
camelot-py==0.11.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
et_xmlfile==2.0.0
fonttools==4.54.1
ghostscript==0.7
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdfminer.six==20240706
pillow==11.0.0
pycparser==2.22
PyMuPDF==1.24.13
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyparsing==3.2.0
pypdf==5.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
screeninfo==0.8.1
setuptools==75.3.0
six==1.16.0
tabulate==0.9.0
tzdata==2024.2
